delete book parameter changed from int to long - delete method simplified

+++ src/main/java/pl/coderslab/BookController.java
@@ -38,7 +38,7 @@ public class BookController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteBook(@PathVariable int id) {
+    public void deleteBook(@PathVariable long id) {
         mockBookService.deleteBook(id);
     }
